// bot/src/foundation/repository/types.ts
export type LogicFilterInfoType = 'and' | 'or'

export type ConditionFilterInfoPredicate =
  | 'eq'
  | 'ge'
  | 'le'
  | 'ne'
  | 'in'
  | 'notin'
  | 'gt'
  | 'lt'
  | 'contains'
  | 'isnull'
  | 'notnull'
  | 'regex'
  | 'startsWith'
  | 'endsWith'
  | 'arrayContains'

export type LogicFilterInfo = {
  type: LogicFilterInfoType
  operands: FilterInfo[]
}

export type ConditionFilterInfo = {
  type: 'condition'
  field: string
  predicate: ConditionFilterInfoPredicate
  value: any
}

export type FilterInfo = LogicFilterInfo | ConditionFilterInfo